third_party/WebKit/LayoutTests/fast/forms/select/listbox-drag-in-from-outside.html

<!DOCTYPE html>
<body>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../resources/common.js"></script>
<div id=log></div>
<select size="4">
<option selected>o1
<option>o2
<option>o3
<option>o4
</select>
<div id="origin">Origin</div>
<script>
var select1 = document.querySelector('select');
var origin = document.querySelector('#origin');

test(function() {
    assert_equals(select1.selectedIndex, 0);
    hoverOverElement(origin);
    eventSender.mouseDown();
    hoverOverElement(select1);
    eventSender.mouseUp();
    assert_equals(select1.selectedIndex, 0);
}, 'Dragging in a listbox from the outside of the listbox should not change selection.');
</script>
</body>
